move tests to attester
